document pattern matcher, add ellipsis groups
Add precise definitions, with examples, for concepts like the ellipsis
nesting level. This should clarify what the matcher is doing. They
should apply to any implementation of the Macrological Fascicle's
description of patterns.

This also adds ellipsis grouping. This is used to determine which
identifiers are allowed to be repeated with each other in output.

TODO: The concept of group needs to encompass nested identifiers.
For instance

    (let-values (((name ...) value) ...) body ...)

does not allow

    ((name ...) body ...)

but the current system does not handle this.